+++ Bsp/tcd1304.h
@@ -60,7 +60,7 @@ typedef enum
 /* Exported variables --------------------------------------------------------*/
 /* Exported functions --------------------------------------------------------*/
 int32_t  TCD_Init(TCD_CONFIG_t *config);
-uint32_t TCD_GetNumOfSpectrumsAcquired(void);
+uint64_t TCD_GetNumOfSpectrumsAcquired(void);
 
 #ifdef __cplusplus
 }
